Splitted HDL ATMegaxx0_1 UART into several files
