+++ hdl/avr/mega/xx0_1/include/uart/regs.hpp
@@ -1,0 +1,36 @@
+#ifndef HDL_AVR_MEGA_XX0_1_UART_REGS_HPP
+#define HDL_AVR_MEGA_XX0_1_UART_REGS_HPP
+
+#include <util/memory_mapping.hpp>
+#include <util/dummy.hpp>
+
+
+namespace HDL {
+    
+  namespace ATMegaxx0_1 {  
+
+    namespace Private {
+
+      namespace UART {
+
+	template<typename ModuleDefs>
+	struct UART_ModuleRegisters {
+	  using Defs = ModuleDefs;
+
+	  static constexpr Register8<typename Defs::Groups::CSR_A> CSR_A = Defs::CSR_A;
+	  static constexpr Register8<typename Defs::Groups::CSR_B> CSR_B = Defs::CSR_B;
+	  static constexpr Register8<typename Defs::Groups::CSR_C> CSR_C = Defs::CSR_C;
+
+	  static constexpr Register8<typename Defs::Groups::BRR> BRR = Defs::BRR;
+	  static constexpr Register8<typename Defs::Groups::BRR_L> BRR_L = Defs::BRR_L;
+	  static constexpr Register8<typename Defs::Groups::BRR_H> BRR_H = Defs::BRR_H;
+
+	  static constexpr Register8<typename Defs::Groups::DR> DR = Defs::DR;
+	};
+
+      }
+    }
+  }
+}
+
+#endif//HDL_AVR_MEGA_XX0_1_UART_REGS_HPP
